reorder the application structure to facilitate the development process
