+++ app/models/__init__.py
@@ -1,0 +1,2 @@
+from .sensor import Sensor
+from .user import User
